finish edit css for detail country page
